Move lib/service/servicecfg.PAMConfig to session/pam/pamcfg.PAMConfig
