+++ session/pam/pam_test.go
@@ -31,7 +31,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/gravitational/teleport/lib/service/servicecfg"
+	"github.com/gravitational/teleport/session/pam/pamcfg"
 )
 
 func TestMain(m *testing.M) {
@@ -72,7 +72,7 @@ func TestEcho(t *testing.T) {
 	username := currentUser(t)
 
 	var buf bytes.Buffer
-	pamContext, err := Open(&servicecfg.PAMConfig{
+	pamContext, err := Open(&pamcfg.PAMConfig{
 		Enabled:     true,
 		ServiceName: "teleport-acct-echo",
 		Login:       username,
@@ -113,7 +113,7 @@ func TestEnvironment(t *testing.T) {
 	username := currentUser(t)
 
 	var buf bytes.Buffer
-	pamContext, err := Open(&servicecfg.PAMConfig{
+	pamContext, err := Open(&pamcfg.PAMConfig{
 		Enabled:     true,
 		ServiceName: "teleport-session-environment",
 		Login:       username,
@@ -133,7 +133,7 @@ func TestSuccess(t *testing.T) {
 	username := currentUser(t)
 
 	var buf bytes.Buffer
-	pamContext, err := Open(&servicecfg.PAMConfig{
+	pamContext, err := Open(&pamcfg.PAMConfig{
 		Enabled:     true,
 		ServiceName: "teleport-success",
 		Login:       username,
@@ -158,7 +158,7 @@ func TestAccountFailure(t *testing.T) {
 	username := currentUser(t)
 
 	var buf bytes.Buffer
-	_, err := Open(&servicecfg.PAMConfig{
+	_, err := Open(&pamcfg.PAMConfig{
 		Enabled:     true,
 		ServiceName: "teleport-acct-failure",
 		Login:       username,
@@ -175,7 +175,7 @@ func TestAuthFailure(t *testing.T) {
 	username := currentUser(t)
 
 	var buf bytes.Buffer
-	_, err := Open(&servicecfg.PAMConfig{
+	_, err := Open(&pamcfg.PAMConfig{
 		Enabled:     true,
 		ServiceName: "teleport-auth-failure",
 		Login:       username,
@@ -193,7 +193,7 @@ func TestAuthDisabled(t *testing.T) {
 	username := currentUser(t)
 
 	var buf bytes.Buffer
-	pamContext, err := Open(&servicecfg.PAMConfig{
+	pamContext, err := Open(&pamcfg.PAMConfig{
 		Enabled:     true,
 		ServiceName: "teleport-auth-failure",
 		Login:       username,
@@ -217,7 +217,7 @@ func TestSessionFailure(t *testing.T) {
 	username := currentUser(t)
 
 	var buf bytes.Buffer
-	_, err := Open(&servicecfg.PAMConfig{
+	_, err := Open(&pamcfg.PAMConfig{
 		Enabled:     true,
 		ServiceName: "teleport-session-failure",
 		Login:       username,
