+++ session/pam/pamcfg/pam.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package servicecfg
+package pamcfg
 
 import (
 	"io"
